refactor: collapse to single sgr classic agent script
