--- .gitignore
@@ -1,351 +1,0 @@
-# SGR Research Agent - .gitignore
-
-# БЕЗОПАСНОСТЬ - API КЛЮЧИ И КОНФИДЕНЦИАЛЬНЫЕ ДАННЫЕ
-config.yaml
-*.env
-.env*
-venv/
-secrets.yaml
-secrets.json
-config.yaml.bak
-config.yaml.backup
-*.key
-*.secret
-
-# PYTHON
-__pycache__/
-*.py[cod]
-*$py.class
-*.so
-.Python
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-wheels/
-share/python-wheels/
-*.egg-info/
-.installed.cfg
-*.egg
-MANIFEST
-
-# PyInstaller
-*.manifest
-*.spec
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.nox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.py,cover
-.hypothesis/
-.pytest_cache/
-cover/
-
-# Translations
-*.mo
-*.pot
-
-# Django stuff:
-*.log
-local_settings.py
-db.sqlite3
-db.sqlite3-journal
-
-# Flask stuff:
-instance/
-.webassets-cache
-
-# Scrapy stuff:
-.scrapy
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-.pybuilder/
-target/
-
-# Jupyter Notebook
-.ipynb_checkpoints
-
-# IPython
-profile_default/
-ipython_config.py
-
-# pyenv
-.python-version
-
-# pipenv
-Pipfile.lock
-
-# poetry
-poetry.lock
-
-# pdm
-.pdm.toml
-
-# PEP 582
-__pypackages__/
-
-# Celery stuff
-celerybeat-schedule
-celerybeat.pid
-
-# SageMath parsed files
-*.sage.py
-
-# Environments
-.env
-.venv
-env/
-venv/
-ENV/
-env.bak/
-venv.bak/
-
-# Spyder project settings
-.spyderproject
-.spyproject
-
-# Rope project settings
-.ropeproject
-
-# mkdocs documentation
-/site
-
-# mypy
-.mypy_cache/
-.dmypy.json
-dmypy.json
-
-# Pyre type checker
-.pyre/
-
-# pytype static type analyzer
-.pytype/
-
-# Cython debug symbols
-cython_debug/
-
-# PyCharm
-.idea/
-
-# ОПЕРАЦИОННЫЕ СИСТЕМЫ
-# Windows
-Thumbs.db
-Thumbs.db:encryptable
-ehthumbs.db
-ehthumbs_vista.db
-*.stackdump
-[Dd]esktop.ini
-$RECYCLE.BIN/
-*.cab
-*.msi
-*.msix
-*.msm
-*.msp
-*.lnk
-
-# macOS
-.DS_Store
-.AppleDouble
-.LSOverride
-Icon
-._*
-.DocumentRevisions-V100
-.fseventsd
-.Spotlight-V100
-.TemporaryItems
-.Trashes
-.VolumeIcon.icns
-.com.apple.timemachine.donotpresent
-.AppleDB
-.AppleDesktop
-Network Trash Folder
-Temporary Items
-.apdisk
-
-# Linux
-*~
-.fuse_hidden*
-.directory
-.Trash-*
-.nfs*
-
-# РЕДАКТОРЫ И IDE
-# Visual Studio Code
-.vscode/
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-!.vscode/*.code-snippets
-
-# Project artifacts and non-essentials
-logs/
-reports/
-.tmp/
-.kiro/
-AGENTS.md
-**/test_*.py
-simple_jazz_research.py
-simple_openai_test.py
-debug_sgr_config.py
-*.patch
-
-# Local History for Visual Studio Code
-.history/
-
-# Built Visual Studio Code Extensions
-*.vsix
-
-# Vim
-*.swp
-*.swo
-*~
-.vimrc.local
-
-# Emacs
-*~
-\#*\#
-/.emacs.desktop
-/.emacs.desktop.lock
-*.elc
-auto-save-list
-tramp
-.\#*
-
-# Sublime Text
-*.sublime-workspace
-*.sublime-project
-
-# Atom
-.atom/
-
-# SGR RESEARCH AGENT СПЕЦИФИЧНЫЕ ФАЙЛЫ
-# Временные файлы исследований
-temp_research/
-tmp/
-.tmp/
-
-# Кэш поисковых результатов
-search_cache/
-.search_cache/
-*.cache
-
-# Логи выполнения
-logs/
-*.log
-execution_*.json
-debug_*.txt
-
-# Отчеты (кроме примера)
-reports/
-!reports/.gitkeep
-
-# Резервные копии отчетов
-*.bak
-*.backup
-temp_*
-draft_*
-
-# Пользовательские данные
-user_data/
-personal/
-private/
-
-# Экспериментальные файлы
-experiments/
-test_runs/
-sandbox/
-
-# РАЗРАБОТКА И ТЕСТИРОВАНИЕ
-# Тестовые файлы
-test_*.py
-*_test.py
-tests/temp/
-test_output/
-
-# Профилирование
-*.prof
-*.profile
-
-# Бенчмарки
-benchmark_results/
-performance_*.json
-
-# Временные скрипты
-temp_*.py
-scratch_*.py
-quick_test_*.py
-
-# ДОКУМЕНТАЦИЯ И МЕДИА
-# Временные файлы документации
-docs/temp/
-*.tmp.md
-draft_*.md
-
-# Большие медиа файлы
-*.mp4
-*.avi
-*.mov
-*.wmv
-*.flv
-*.webm
-
-# Архивы
-*.zip
-*.tar.gz
-*.rar
-*.7z
-
-# СЕТЕВЫЕ И СИСТЕМНЫЕ ФАЙЛЫ
-# Сертификаты
-*.pem
-*.crt
-*.p12
-*.pfx
-
-# SSH ключи
-id_rsa
-id_rsa.pub
-id_ed25519
-id_ed25519.pub
-
-# РАЗНОЕ
-# Файлы блокировки
-*.lock
-.lock
-
-# Временные файлы
-*.tmp
-*.temp
-
-# Пользовательские заметки
-notes.txt
-todo.txt
-personal_notes.md
-
-# ВАЖНО: config.yaml.example должен быть в репозитории как шаблон
-!config.yaml.example
-
-# Conversation logs for debugging
-logs/
