--- example_report.md
@@ -1,33 +1,0 @@
-# Что такое SGR и его применение в LLM
-
-*Created: 2025-08-26 05:37:14*
-
-### Исполнительное резюме
-Schema Guided Reasoning (SGR) — это новая техника рассуждения, используемая в больших языковых моделях (LLM), которая позволяет адаптировать логику под каждую конкретную задачу. Это улучшает производительность моделей, позволяя им использовать специализированные схемы рассуждений, что особенно важно для контекстно-осведомленных приложений искусственного интеллекта.
-
-### Технический анализ
-SGR представляет собой подход, который позволяет моделям LLM использовать заранее определенные схемы рассуждений для решения задач. Это значит, что вместо применения универсального подхода, SGR позволяет создавать индивидуализированные логические структуры, которые лучше соответствуют специфике каждой задачи. Например, в новых моделях OpenAI SGR показал рекордные результаты благодаря своей способности адаптироваться к различным задачам [1].
-
-SGR включает в себя использование простых графовых паттернов, таких как Cascade, Routing и Cycle, которые помогают в построении сложных схем рассуждений и в генерации синтетических данных [6]. Эти паттерны позволяют моделям лучше понимать зависимости между данными и адаптироваться к изменяющимся условиям задачи.
-
-Применение SGR в LLM также включает в себя генерацию SQL-запросов, где модели могут рекомендовать таблицы и условия JOIN на основе предыдущего опыта и текущих потребностей [8]. Это снижает нагрузку на аналитиков и минимизирует риски, связанные с человеческим фактором.
-
-### Ключевые выводы
-1. SGR — это новая техника, улучшающая производительность LLM через адаптацию логики под конкретные задачи.
-2. Использование графовых паттернов позволяет моделям более эффективно обрабатывать данные и строить зависимости.
-3. Применение SGR в различных областях, таких как генерация SQL-запросов, демонстрирует его универсальность и эффективность.
-
-### Заключение
-SGR является важным шагом вперед в развитии LLM, позволяя моделям более эффективно справляться с задачами, требующими сложного рассуждения и анализа данных. Эта технология открывает новые возможности для создания контекстно-осведомленных приложений ИИ, которые могут адаптироваться к специфическим требованиям пользователей и задач.
-
-## Sources
-[1] Новые модели OpenAI показали рекордные результаты ... - https://fittin.ru/contentzavod/view/razrabotka-mobilnyh-prilozheniy-pod-klyuch-novye-modeli-openai-pokazali-rekordnye-rezultaty-v-sgr
-[2] Анатомия памяти LLM: Почему будущее не за ... - https://habr.com/ru/articles/934244/
-[3] Промпт-инжиниринг на основе здравого смысла - https://habr.com/ru/articles/902230/
-[4] Публикация #1579 — Neural Deep (@neuraldeep) - https://tgstat.ru/channel/@neuraldeep/1579
-[5] Построение LLM инфраструктуры и создание AI ... - https://www.youtube.com/watch?v=7Njad-XzHr4
-[6] Генерация синтетических данных для LLM. Часть 2: графовый ... - https://habr.com/ru/companies/sberbank/articles/909934/
-[7] Как лучше всего сделать так, чтобы все мои персонажи узнавали ... - https://www.reddit.com/r/SillyTavernAI/comments/1m8cxft/how_best_should_i_go_about_getting_all_my/?tl=ru
-[8] Как LLM могут помочь аналитикам баз данных в работе с SQL ... - https://habr.com/ru/companies/sberbank/articles/909730/
-[9] Разработка простого LLM-приложения - Sber Developers - https://developers.sber.ru/docs/ru/gigachain/tutorials/llm-chain
-[10] Telegram channel "LLM под капотом" — @llm_under_hood — TGStat - https://tgstat.com/channel/%40llm_under_hood
