--- sgr-classic/config.yaml.example
@@ -1,55 +1,0 @@
-# SGR Research Agent - Configuration Template
-# Production-ready configuration for Schema-Guided Reasoning
-# Скопируйте этот файл в config.yaml и заполните своими API ключами
-
-# OpenAI Configuration
-openai:
-  api_key: "your-openai-api-key-here"  # Обязательно: Ваш OpenAI API ключ
-  base_url: ""                         # Опционально: Альтернативный URL (например, для прокси)
-  model: "gpt-4o-mini"                 # Модель для использования
-  max_tokens: 8000                     # Максимальное количество токенов
-  temperature: 0.4                     # Температура генерации (0.0-1.0)
-
-# Tavily Search Configuration
-tavily:
-  api_key: "your-tavily-api-key-here"  # Обязательно: Ваш Tavily API ключ
-
-# Search Settings
-search:
-  max_results: 10                      # Максимальное количество результатов поиска
-
-# Scraping Settings
-scraping:
-  enabled: false                       # Включить скрейпинг полного текста найденных страниц
-  max_pages: 5                         # Максимум страниц для скрейпинга за поиск
-  content_limit: 1500                  # Лимит символов полного контента на источник
-
-# Execution Settings
-execution:
-  max_steps: 6                         # Максимальное количество шагов выполнения
-  reports_dir: "reports"               # Директория для сохранения отчетов
-
-
-
-# ============================================================================
-# ИНСТРУКЦИИ ПО НАСТРОЙКЕ:
-# ============================================================================
-#
-# 1. Получите OpenAI API ключ:
-#    - Зайдите на https://platform.openai.com/api-keys
-#    - Создайте новый API ключ
-#    - Скопируйте его в поле openai.api_key
-#
-# 2. Получите Tavily API ключ:
-#    - Зайдите на https://tavily.com
-#    - Зарегистрируйтесь и получите API ключ
-#    - Скопируйте его в поле tavily.api_key
-#
-# 3. Альтернативно, можете использовать переменные окружения:
-#    export OPENAI_API_KEY="your-openai-key"
-#    export TAVILY_API_KEY="your-tavily-key"
-#
-# 4. Создайте директорию для отчетов:
-#    mkdir reports
-#
-# ============================================================================
