--- sgr-classic/requirements.txt
@@ -1,18 +1,0 @@
-# SGR Classic Requirements
-# Базовые зависимости для классической версии SGR
-
-# Core dependencies
-openai>=1.0.0
-pydantic>=2.0.0
-rich>=13.0.0
-
-# Search and research
-tavily-python>=0.3.0
-
-# Web scraping and content extraction
-trafilatura>=1.6.0
-youtube-transcript-api>=0.6.0
-
-# Configuration and utilities
-PyYAML>=6.0
-python-dateutil>=2.8.0
