--- sgr-streaming/requirements.txt
@@ -1,32 +1,0 @@
-# SGR Streaming Requirements  
-# Расширенные зависимости для версии с стримингом
-
-# Core dependencies (same as classic)
-openai>=1.0.0
-pydantic>=2.0.0
-rich>=13.0.0
-
-# Enhanced typing for streaming
-annotated-types>=0.6.0
-typing-extensions>=4.8.0
-
-# Search and research
-tavily-python>=0.3.0
-
-# Web scraping and content extraction
-trafilatura>=1.6.0
-youtube-transcript-api>=0.6.0
-
-# Configuration and utilities
-PyYAML>=6.0
-python-dateutil>=2.8.0
-requests>=2.31.0
-
-# LiteLLM proxy (front of Ollama)
-litellm>=1.35.0
-fastapi>=0.111.0
-uvicorn[standard]>=0.30.0
-httpx>=0.27.0
-Jinja2>=3.1.0
-opentelemetry-api>=1.24.0
-opentelemetry-sdk>=1.24.0
