--- sgr-streaming/schema_enforcement_engine.py
@@ -1,805 +1,0 @@
-#!/usr/bin/env python3
-"""
-Schema Enforcement Engine
-Comprehensive validation and correction for all SGR schema issues
-"""
-
-from typing import Dict, List, Any, Union
-import re
-import json
-
-class SchemaEnforcementEngine:
-    """Proactive schema validation and correction engine with logic gates"""
-    
-    def __init__(self):
-        self.tool_name_corrections = {
-            'clarify': 'clarification',
-            'plan': 'generate_plan', 
-            'search': 'web_search',
-            'adapt': 'adapt_plan',
-            'report': 'create_report',
-            'complete': 'report_completion',
-            'completion': 'report_completion'
-        }
-        
-        self.default_content_generators = {
-            'clarification': self._generate_clarification_defaults,
-            'generate_plan': self._generate_plan_defaults,
-            'web_search': self._generate_search_defaults,
-            'adapt_plan': self._generate_adapt_defaults,
-            'create_report': self._generate_report_defaults,
-            'report_completion': self._generate_completion_defaults
-        }
-        
-        # Logic gates for workflow progression
-        self.workflow_logic_gates = {
-            'no_plan_no_searches': {
-                'condition': lambda ctx: not ctx.get('has_plan') and ctx.get('searches_done', 0) == 0,
-                'allowed_tools': ['clarification', 'generate_plan'],
-                'forced_tool': 'generate_plan',  # Force progression
-                'reason': 'No plan exists, must create plan first'
-            },
-            'plan_exists_no_searches': {
-                'condition': lambda ctx: ctx.get('has_plan') and ctx.get('searches_done', 0) == 0,
-                'allowed_tools': ['web_search'],
-                'forced_tool': 'web_search',
-                'reason': 'Plan exists but no searches done, must start searching'
-            },
-            'plan_exists_few_searches': {
-                'condition': lambda ctx: ctx.get('has_plan') and 0 < ctx.get('searches_done', 0) < 2,
-                'allowed_tools': ['web_search', 'adapt_plan'],
-                'forced_tool': 'web_search',  # Continue searching unless adaptation needed
-                'reason': 'Plan exists with few searches, continue gathering data'
-            },
-            'sufficient_searches': {
-                'condition': lambda ctx: ctx.get('searches_done', 0) >= 2,
-                'allowed_tools': ['create_report', 'web_search', 'adapt_plan'],
-                'forced_tool': 'create_report',  # Force report creation after sufficient searches
-                'reason': 'Sufficient searches completed, must create report'
-            },
-            'empty_arrays_exit': {
-                'condition': lambda ctx: (
-                    not ctx.get('remaining_steps') or 
-                    all(not step.strip() for step in ctx.get('remaining_steps', [])) or
-                    ctx.get('remaining_steps') == ['']
-                ),
-                'allowed_tools': ['create_report', 'report_completion'],
-                'forced_tool': 'create_report',
-                'reason': 'Empty remaining steps array, must complete task'
-            }
-        }
-    
-    def enforce_schema(self, data: Dict[str, Any], context: Dict[str, Any] = None) -> Dict[str, Any]:
-        """Main entry point - enforce all schema requirements with logic gates"""
-        if not isinstance(data, dict):
-            return self._create_fallback_clarification()
-        
-        context = context or {}
-        
-        # Step 1: Anti-clarification-loop check
-        data = self._prevent_clarification_loops(data, context)
-        
-        # Step 2: Fix basic data types
-        data = self._fix_data_types(data)
-        
-        # Step 3: Ensure NextStep structure
-        data = self._ensure_nextstep_structure(data)
-        
-        # Step 4: LOGIC GATES - Enforce workflow progression rules
-        data = self._enforce_logic_gates(data, context)
-        
-        # Step 5: Fix function object (after logic gates)
-        data = self._fix_function_object(data)
-        
-        # Step 6: Validate and fix all lists
-        data = self._fix_all_lists(data)
-        
-        # Step 7: Fix empty strings
-        data = self._fix_empty_strings(data)
-        
-        # Step 8: Business logic validation
-        data = self._validate_business_logic(data)
-        
-        # Step 9: Final validation - ensure no broken patterns
-        data = self._final_pattern_validation(data, context)
-        
-        return data
-    
-    def _enforce_logic_gates(self, data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Enforce workflow logic gates to prevent broken patterns"""
-        
-        # Build workflow context
-        workflow_context = self._build_workflow_context(data, context)
-        
-        # Get current tool
-        current_tool = data.get('function', {}).get('tool', 'clarification')
-        
-        print(f"🚪 Logic Gates: Current tool={current_tool}, Context={workflow_context}")
-        
-        # Check each logic gate
-        for gate_name, gate_config in self.workflow_logic_gates.items():
-            if gate_config['condition'](workflow_context):
-                print(f"🚪 Logic Gate TRIGGERED: {gate_name}")
-                print(f"🚪 Reason: {gate_config['reason']}")
-                
-                # Check if current tool is allowed
-                if current_tool not in gate_config['allowed_tools']:
-                    print(f"🚪 BLOCKING: {current_tool} not in allowed tools {gate_config['allowed_tools']}")
-                    
-                    # Force the correct tool
-                    forced_tool = gate_config['forced_tool']
-                    print(f"🚪 FORCING: {current_tool} -> {forced_tool}")
-                    
-                    # Update the function tool
-                    if 'function' not in data:
-                        data['function'] = {}
-                    data['function']['tool'] = forced_tool
-                    
-                    # Update reasoning to explain the forced transition
-                    data['function']['reasoning'] = f"Logic gate enforcement: {gate_config['reason']}"
-                    
-                    # Update current situation
-                    data['current_situation'] = f"Workflow progression enforced: {gate_config['reason']}"
-                    
-                    # Update reasoning steps
-                    data['reasoning_steps'] = [
-                        f"Logic gate triggered: {gate_name}",
-                        f"Enforcing workflow progression: {current_tool} -> {forced_tool}"
-                    ]
-                    
-                    break  # Only apply first matching gate
-                else:
-                    print(f"🚪 ALLOWED: {current_tool} is in allowed tools {gate_config['allowed_tools']}")
-        
-        return data
-    
-    def _build_workflow_context(self, data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Build workflow context for logic gate evaluation"""
-        
-        # Check if plan exists
-        has_plan = (
-            data.get('plan_status', '').lower() == 'active' or
-            context.get('plan') is not None or
-            'plan' in data.get('current_situation', '').lower() or
-            any('plan' in step.lower() for step in data.get('reasoning_steps', []))
-        )
-        
-        # Get searches done
-        searches_done = data.get('searches_done', 0)
-        
-        # Check remaining steps
-        remaining_steps = data.get('remaining_steps', [])
-        has_remaining_steps = (
-            remaining_steps and 
-            len(remaining_steps) > 0 and
-            not all(not step.strip() for step in remaining_steps) and
-            remaining_steps != ['']
-        )
-        
-        # Check if enough data
-        enough_data = data.get('enough_data', False)
-        
-        # Check task completion
-        task_completed = data.get('task_completed', False)
-        
-        workflow_context = {
-            'has_plan': has_plan,
-            'searches_done': searches_done,
-            'has_remaining_steps': has_remaining_steps,
-            'enough_data': enough_data,
-            'task_completed': task_completed,
-            'remaining_steps': remaining_steps
-        }
-        
-        print(f"🔍 Workflow Context: {workflow_context}")
-        return workflow_context
-    
-    def _final_pattern_validation(self, data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Final validation to catch any remaining broken patterns"""
-        
-        current_tool = data.get('function', {}).get('tool', 'clarification')
-        searches_done = data.get('searches_done', 0)
-        plan_status = data.get('plan_status', '').lower()
-        remaining_steps = data.get('remaining_steps', [])
-        
-        # Pattern 1: Empty remaining steps should trigger completion
-        if (not remaining_steps or 
-            all(not step.strip() for step in remaining_steps) or 
-            remaining_steps == ['']):
-            
-            if current_tool not in ['create_report', 'report_completion']:
-                print("🔒 FINAL VALIDATION: Empty remaining steps, forcing create_report")
-                data['function']['tool'] = 'create_report'
-                data['task_completed'] = True
-                data['current_situation'] = 'No remaining steps, completing task'
-        
-        # Pattern 2: Active plan with no searches should search
-        elif plan_status == 'active' and searches_done == 0 and current_tool == 'generate_plan':
-            print("🔒 FINAL VALIDATION: Active plan with no searches, forcing web_search")
-            data['function']['tool'] = 'web_search'
-            data['current_situation'] = 'Plan active, beginning search phase'
-        
-        # Pattern 3: Multiple searches done should create report
-        elif searches_done >= 2 and current_tool in ['generate_plan', 'web_search']:
-            print("🔒 FINAL VALIDATION: Sufficient searches done, forcing create_report")
-            data['function']['tool'] = 'create_report'
-            data['enough_data'] = True
-            data['current_situation'] = 'Sufficient data gathered, creating report'
-        
-        return data
-    
-    def _prevent_clarification_loops(self, data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Prevent endless clarification loops - force action when appropriate"""
-        
-        # Check if user has given clear direction to proceed
-        user_messages = context.get('user_messages', [])
-        proceed_keywords = ['begin', 'start', 'proceed', 'go ahead', 'continue', 'do it', 'yes']
-        
-        user_wants_to_proceed = False
-        if user_messages:
-            latest_message = str(user_messages[-1]).lower()
-            user_wants_to_proceed = any(keyword in latest_message for keyword in proceed_keywords)
-        
-        # Check if we've already clarified before
-        clarification_used = context.get('clarification_used', False)
-        
-        # ONLY prevent clarification loops - don't interfere with other tools
-        func = data.get('function', {})
-        current_tool = func.get('tool')
-        
-        # Only intervene if it's trying to clarify when it shouldn't
-        if current_tool == 'clarification' and (user_wants_to_proceed or clarification_used):
-            print("🚨 Preventing clarification loop - forcing generate_plan")
-            
-            # Convert to generate_plan
-            data['function'] = {
-                'tool': 'generate_plan',
-                'reasoning': 'User provided direction to proceed with research',
-                'research_goal': 'Conduct research based on user request',
-                'planned_steps': [
-                    'Research background information',
-                    'Gather comprehensive data',
-                    'Analyze and synthesize findings'
-                ],
-                'search_strategies': [
-                    'Web search for reliable sources',
-                    'Verify information accuracy'
-                ]
-            }
-            
-            # Update other fields accordingly
-            data['current_situation'] = 'Ready to begin research based on user direction'
-            data['plan_status'] = 'active'
-            data['reasoning_steps'] = [
-                'User provided clear direction to proceed',
-                'Generating research plan to begin work'
-            ]
-        
-        return data
-    
-    def _fix_data_types(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Fix incorrect data types"""
-        # Convert string numbers to int
-        if 'searches_done' in data:
-            try:
-                data['searches_done'] = int(data['searches_done'])
-            except (ValueError, TypeError):
-                data['searches_done'] = 0
-        
-        # Convert string booleans to bool
-        for bool_field in ['enough_data', 'task_completed', 'plan_adapted', 'scrape_content']:
-            if bool_field in data:
-                val = data[bool_field]
-                if isinstance(val, str):
-                    data[bool_field] = val.lower() in ('true', '1', 'yes')
-                elif not isinstance(val, bool):
-                    data[bool_field] = bool(val)
-        
-        # Fix remaining_steps if it contains objects
-        if 'remaining_steps' in data and isinstance(data['remaining_steps'], list):
-            fixed_steps = []
-            for step in data['remaining_steps']:
-                if isinstance(step, dict):
-                    # Extract meaningful string from object
-                    step_str = (step.get('step_name') or 
-                               step.get('name') or 
-                               step.get('description') or
-                               step.get('action_type', '').replace('_', ' ').title() or
-                               f"Execute {step.get('function', {}).get('tool', 'action')}")
-                    
-                    # Add query if present
-                    if 'search_query' in step:
-                        step_str += f": {step['search_query']}"
-                    
-                    fixed_steps.append(str(step_str))
-                else:
-                    fixed_steps.append(str(step))
-            data['remaining_steps'] = fixed_steps
-        
-        return data
-    
-    def _ensure_nextstep_structure(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Ensure all required NextStep fields exist"""
-        
-        # Infer function type from context if missing
-        inferred_function = self._infer_function_from_context(data)
-        
-        defaults = {
-            'reasoning_steps': ["Processing user request", "Determining next action"],
-            'current_situation': 'Processing user request',
-            'plan_status': 'active',
-            'searches_done': 0,
-            'enough_data': False,
-            'remaining_steps': ['Continue with plan'],
-            'task_completed': False,
-            'function': inferred_function
-        }
-        
-        for field, default_value in defaults.items():
-            if field not in data or data[field] is None:
-                data[field] = default_value
-        
-        return data
-    
-    def _infer_function_from_context(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Infer the correct function type from context clues"""
-        
-        # Check reasoning steps and remaining steps for clues
-        reasoning_text = ' '.join(data.get('reasoning_steps', [])).lower()
-        remaining_text = str(data.get('remaining_steps', [])).lower()
-        plan_status = data.get('plan_status', '').lower()
-        searches_done = data.get('searches_done', 0)
-        
-        # If we have an active plan and no searches yet, we should search
-        if plan_status == 'active' and searches_done == 0:
-            # Look for search indicators first
-            search_keywords = ['web search', 'search for', 'find information', 'gather data']
-            if any(keyword in reasoning_text or keyword in remaining_text for keyword in search_keywords):
-                return {
-                    'tool': 'web_search',
-                    'reasoning': 'Executing planned web search to gather information',
-                    'query': 'research topic information',
-                    'max_results': 10,
-                    'plan_adapted': False,
-                    'scrape_content': False
-                }
-        
-        # If we have searches done but not enough data, continue searching
-        if searches_done > 0 and searches_done < 3 and not data.get('enough_data', False):
-            search_keywords = ['web search', 'search for', 'find information', 'gather data']
-            if any(keyword in reasoning_text or keyword in remaining_text for keyword in search_keywords):
-                return {
-                    'tool': 'web_search',
-                    'reasoning': 'Continuing research to gather more comprehensive information',
-                    'query': 'additional research information',
-                    'max_results': 10,
-                    'plan_adapted': False,
-                    'scrape_content': False
-                }
-        
-        # If we have enough data, create report
-        if data.get('enough_data', False) or searches_done >= 2:
-            report_keywords = ['create report', 'report', 'summarize', 'compile findings']
-            if any(keyword in reasoning_text or keyword in remaining_text for keyword in report_keywords):
-                return {
-                    'tool': 'create_report',
-                    'reasoning': 'Ready to create comprehensive report based on gathered information',
-                    'title': 'Research Report',
-                    'user_request_language_reference': 'research request',
-                    'content': 'Comprehensive research report based on gathered information.',
-                    'confidence': 'medium'
-                }
-        
-        # Look for explicit planning indicators
-        planning_keywords = ['generate plan', 'plan for', 'research plan', 'planning', 'strategy']
-        has_planning_keywords = any(keyword in reasoning_text for keyword in planning_keywords)
-        
-        # Only infer generate_plan if there are explicit planning keywords AND no active plan
-        if has_planning_keywords and plan_status != 'active':
-            return {
-                'tool': 'generate_plan',
-                'reasoning': 'Generated research plan based on user request',
-                'research_goal': 'Conduct comprehensive research on the requested topic',
-                'planned_steps': [
-                    'Research background information',
-                    'Gather comprehensive data', 
-                    'Analyze and synthesize findings'
-                ],
-                'search_strategies': [
-                    'Web search for current information',
-                    'Verify source credibility'
-                ]
-            }
-        
-        # Default to clarification only if nothing else makes sense
-        return {
-            'tool': 'clarification',
-            'reasoning': 'Need clarification to proceed',
-            'questions': [
-                'Could you provide more specific details?',
-                'What particular aspect interests you?',
-                'Are there specific requirements to consider?'
-            ],
-            'unclear_terms': ['Request scope'],
-            'assumptions': [
-                'You want comprehensive information',
-                'Current data is preferred'
-            ]
-        }
-    
-    def _fix_function_object(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Fix and complete function object (logic gates already applied)"""
-        if 'function' not in data or not isinstance(data['function'], dict):
-            data['function'] = {'tool': 'clarification'}
-        
-        func = data['function']
-        
-        # Fix tool name
-        tool = func.get('tool', 'clarification')
-        if tool in self.tool_name_corrections:
-            tool = self.tool_name_corrections[tool]
-            func['tool'] = tool
-        
-        # Generate missing fields based on tool type (logic gates have already corrected the tool)
-        if tool in self.default_content_generators:
-            func = self.default_content_generators[tool](func, data)
-        else:
-            # Unknown tool - default to clarification
-            func = self._generate_clarification_defaults(func, data)
-        
-        data['function'] = func
-        return data
-    
-    def _fix_all_lists(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Fix all list fields to meet minimum length requirements"""
-        list_requirements = {
-            'reasoning_steps': {'min': 2, 'defaults': ["Processing request", "Determining action"]},
-            'remaining_steps': {'min': 1, 'defaults': ["Continue with plan"]},
-        }
-        
-        for field, req in list_requirements.items():
-            if field in data:
-                data[field] = self._ensure_list_min_length(
-                    data[field], req['min'], req['defaults']
-                )
-        
-        # Fix function-specific lists
-        if 'function' in data and isinstance(data['function'], dict):
-            func = data['function']
-            tool = func.get('tool')
-            
-            if tool == 'clarification':
-                func['questions'] = self._ensure_list_min_length(
-                    func.get('questions', []), 3, [
-                        "Could you provide more specific details?",
-                        "What particular aspect interests you most?", 
-                        "Are there specific requirements to consider?"
-                    ]
-                )
-                func['unclear_terms'] = self._ensure_list_min_length(
-                    func.get('unclear_terms', []), 1, ["Request scope"]
-                )
-                func['assumptions'] = self._ensure_list_min_length(
-                    func.get('assumptions', []), 2, [
-                        "You want comprehensive information",
-                        "Current data is preferred"
-                    ]
-                )
-            
-            elif tool == 'generate_plan':
-                func['planned_steps'] = self._ensure_list_min_length(
-                    func.get('planned_steps', []), 3, [
-                        "Research background information",
-                        "Gather comprehensive data",
-                        "Analyze and synthesize findings"
-                    ]
-                )
-                func['search_strategies'] = self._ensure_list_min_length(
-                    func.get('search_strategies', []), 2, [
-                        "Web search for current information",
-                        "Verify sources and credibility"
-                    ]
-                )
-            
-            elif tool == 'adapt_plan':
-                func['plan_changes'] = self._ensure_list_min_length(
-                    func.get('plan_changes', []), 1, ["Updated research approach"]
-                )
-                func['next_steps'] = self._ensure_list_min_length(
-                    func.get('next_steps', []), 2, [
-                        "Continue with adapted plan",
-                        "Gather additional information"
-                    ]
-                )
-            
-            elif tool == 'report_completion':
-                func['completed_steps'] = self._ensure_list_min_length(
-                    func.get('completed_steps', []), 1, ["Completed research task"]
-                )
-        
-        return data
-    
-    def _ensure_list_min_length(self, items: Any, min_len: int, defaults: List[str]) -> List[str]:
-        """Ensure list meets minimum length requirement and convert complex objects to simple strings"""
-        if not isinstance(items, list):
-            items = [str(items)] if items else []
-        
-        # Convert all items to simple strings, handling complex objects
-        cleaned_items = []
-        for item in items:
-            if not item:
-                continue
-                
-            if isinstance(item, dict):
-                # Extract meaningful content from dict objects
-                if 'action' in item and 'query' in item:
-                    # Pattern: {"action": "web_search", "query": "jazz origins"}
-                    action = item['action'].replace('_', ' ').title()
-                    query = item['query']
-                    cleaned_items.append(f"{action}: {query}")
-                elif 'step_name' in item:
-                    cleaned_items.append(str(item['step_name']))
-                elif 'description' in item:
-                    cleaned_items.append(str(item['description']))
-                else:
-                    # Generic dict handling
-                    cleaned_items.append(f"Execute {item.get('action', 'action')}")
-            elif isinstance(item, str):
-                # Handle JSON strings embedded in strings
-                if item.startswith('{') and item.endswith('}'):
-                    try:
-                        parsed = json.loads(item)
-                        if isinstance(parsed, dict):
-                            if 'action' in parsed and 'query' in parsed:
-                                action = parsed['action'].replace('_', ' ').title()
-                                query = parsed['query']
-                                cleaned_items.append(f"{action}: {query}")
-                            elif 'action' in parsed:
-                                action = parsed['action'].replace('_', ' ').title()
-                                cleaned_items.append(f"{action}")
-                            else:
-                                cleaned_items.append("Research step")
-                        else:
-                            cleaned_items.append(str(item))
-                    except:
-                        # If JSON parsing fails, use as-is
-                        cleaned_items.append(str(item))
-                else:
-                    cleaned_items.append(str(item))
-            else:
-                cleaned_items.append(str(item))
-        
-        # Pad with defaults if needed
-        while len(cleaned_items) < min_len:
-            needed = min_len - len(cleaned_items)
-            cleaned_items.extend(defaults[:needed])
-        
-        return cleaned_items[:min_len + 2]  # Cap at reasonable maximum
-    
-    def _fix_empty_strings(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Fix empty required string fields"""
-        string_defaults = {
-            'current_situation': 'Processing user request',
-            'plan_status': 'active'
-        }
-        
-        for field, default in string_defaults.items():
-            if field in data and (not data[field] or data[field].strip() == ''):
-                data[field] = default
-        
-        # Fix function string fields
-        if 'function' in data and isinstance(data['function'], dict):
-            func = data['function']
-            tool = func.get('tool', 'clarification')
-            
-            if not func.get('reasoning') or func['reasoning'].strip() == '':
-                func['reasoning'] = f"Executing {tool.replace('_', ' ')} step"
-            
-            # Tool-specific string fixes
-            if tool == 'generate_plan':
-                if not func.get('research_goal') or func['research_goal'].strip() == '':
-                    func['research_goal'] = "Conduct comprehensive research on the requested topic"
-            
-            elif tool == 'web_search':
-                if not func.get('query') or func['query'].strip() == '':
-                    func['query'] = "research topic information"
-            
-            elif tool == 'create_report':
-                if not func.get('title') or func['title'].strip() == '':
-                    func['title'] = "Research Report"
-                if not func.get('user_request_language_reference') or func['user_request_language_reference'].strip() == '':
-                    func['user_request_language_reference'] = "research request"
-                if not func.get('content') or func['content'].strip() == '':
-                    func['content'] = self._generate_minimum_report_content()
-        
-        return data
-    
-    def _validate_business_logic(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        """Validate business logic and fix critical issues"""
-        func = data.get('function', {})
-        tool = func.get('tool')
-        
-        # Critical: Generated plan with no steps
-        if tool == 'generate_plan':
-            planned_steps = func.get('planned_steps', [])
-            if not planned_steps or all(not step.strip() for step in planned_steps):
-                func['planned_steps'] = [
-                    "Research background and context",
-                    "Gather comprehensive information", 
-                    "Analyze findings and create report"
-                ]
-                print("⚠️ Fixed: Generated plan had no steps")
-        
-        # Critical: Search with empty query
-        if tool == 'web_search':
-            query = func.get('query', '').strip()
-            if not query:
-                # Try to infer from context
-                reasoning_steps = data.get('reasoning_steps', [])
-                if reasoning_steps:
-                    # Extract topic from reasoning
-                    reasoning_text = ' '.join(reasoning_steps).lower()
-                    if 'jazz' in reasoning_text:
-                        func['query'] = 'jazz history research'
-                    else:
-                        func['query'] = 'research topic information'
-                else:
-                    func['query'] = 'research information'
-                print("⚠️ Fixed: Search had empty query")
-        
-        # Critical: Report with insufficient content
-        if tool == 'create_report':
-            content = func.get('content', '').strip()
-            if len(content) < 100:  # Much too short
-                func['content'] = self._generate_minimum_report_content()
-                print("⚠️ Fixed: Report content was too short")
-        
-        return data
-    
-    def _generate_minimum_report_content(self) -> str:
-        """Generate minimum viable report content"""
-        return """# Research Report
-
-## Executive Summary
-This report presents the findings from comprehensive research conducted on the requested topic. The research involved systematic information gathering and analysis to provide accurate and relevant insights.
-
-## Methodology
-The research approach included:
-- Web-based information gathering from credible sources
-- Analysis of current and historical data
-- Synthesis of findings from multiple perspectives
-
-## Key Findings
-Based on the research conducted, several important points have been identified:
-
-1. **Background Context**: The topic has significant historical and contemporary relevance
-2. **Current Status**: Recent developments show ongoing evolution in this area
-3. **Key Factors**: Multiple elements contribute to the current understanding
-
-## Analysis
-The research reveals important patterns and trends that provide insight into the topic. The information gathered demonstrates the complexity and multifaceted nature of the subject matter.
-
-## Conclusions
-The research provides a comprehensive overview of the requested topic. The findings contribute to a better understanding of the key aspects and their implications.
-
-## Sources
-Research conducted using credible web sources and current information available at the time of analysis.
-
-*Note: This report represents findings based on available information and research conducted.*"""
-    
-    # Default generators for each tool type
-    def _generate_clarification_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete clarification function"""
-        return {
-            'tool': 'clarification',
-            'reasoning': func.get('reasoning', 'Need clarification to provide accurate research'),
-            'questions': func.get('questions', [
-                "Could you provide more specific details about your request?",
-                "What particular aspect are you most interested in?",
-                "Are there any specific requirements or constraints I should know about?"
-            ]),
-            'unclear_terms': func.get('unclear_terms', ['Request scope and focus']),
-            'assumptions': func.get('assumptions', [
-                'You want comprehensive and accurate information',
-                'Current and reliable data is preferred'
-            ])
-        }
-    
-    def _generate_plan_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete plan function"""
-        # Try to infer topic from context
-        topic = "the requested topic"
-        reasoning_steps = context.get('reasoning_steps', [])
-        if reasoning_steps:
-            reasoning_text = ' '.join(reasoning_steps).lower()
-            if 'jazz' in reasoning_text:
-                topic = "jazz history"
-        
-        return {
-            'tool': 'generate_plan',
-            'reasoning': func.get('reasoning', f'Generated comprehensive research plan for {topic}'),
-            'research_goal': func.get('research_goal', f'Conduct thorough research on {topic}'),
-            'planned_steps': func.get('planned_steps', [
-                f'Research background and origins of {topic}',
-                f'Investigate key developments and milestones',
-                f'Analyze current status and significance'
-            ]),
-            'search_strategies': func.get('search_strategies', [
-                'Web search for authoritative sources',
-                'Verify information credibility and accuracy'
-            ])
-        }
-    
-    def _generate_search_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete search function"""
-        return {
-            'tool': 'web_search',
-            'reasoning': func.get('reasoning', 'Searching for comprehensive information'),
-            'query': func.get('query', 'research topic information'),
-            'max_results': func.get('max_results', 10),
-            'plan_adapted': func.get('plan_adapted', False),
-            'scrape_content': func.get('scrape_content', False)
-        }
-    
-    def _generate_adapt_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete adapt plan function"""
-        return {
-            'tool': 'adapt_plan',
-            'reasoning': func.get('reasoning', 'Adapting plan based on new findings'),
-            'original_goal': func.get('original_goal', 'Original research objective'),
-            'new_goal': func.get('new_goal', 'Updated research objective'),
-            'plan_changes': func.get('plan_changes', ['Updated research approach']),
-            'next_steps': func.get('next_steps', [
-                'Continue with adapted plan',
-                'Gather additional information'
-            ])
-        }
-    
-    def _generate_report_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete report function"""
-        return {
-            'tool': 'create_report',
-            'reasoning': func.get('reasoning', 'Ready to create comprehensive report'),
-            'title': func.get('title', 'Research Report'),
-            'user_request_language_reference': func.get('user_request_language_reference', 'research request'),
-            'content': func.get('content', self._generate_minimum_report_content()),
-            'confidence': func.get('confidence', 'medium')
-        }
-    
-    def _generate_completion_defaults(self, func: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
-        """Generate complete completion function"""
-        return {
-            'tool': 'report_completion',
-            'reasoning': func.get('reasoning', 'Research task completed successfully'),
-            'completed_steps': func.get('completed_steps', ['Completed comprehensive research']),
-            'status': func.get('status', 'completed')
-        }
-    
-    def _create_fallback_clarification(self) -> Dict[str, Any]:
-        """Create fallback clarification when data is completely invalid"""
-        return {
-            'reasoning_steps': [
-                'Unable to parse user request properly',
-                'Need clarification to proceed with research'
-            ],
-            'current_situation': 'Awaiting user clarification due to parsing issues',
-            'plan_status': 'paused',
-            'searches_done': 0,
-            'enough_data': False,
-            'remaining_steps': ['Receive clarification', 'Generate research plan'],
-            'task_completed': False,
-            'function': {
-                'tool': 'clarification',
-                'reasoning': 'Unable to understand request, need clarification',
-                'questions': [
-                    'Could you please rephrase your request more clearly?',
-                    'What specific information are you looking for?',
-                    'Are there particular aspects you want me to focus on?'
-                ],
-                'unclear_terms': ['Request format and content'],
-                'assumptions': [
-                    'You want me to conduct research',
-                    'Clear communication will help provide better results'
-                ]
-            }
-        }
